refactor ♻️: Rename and refactor CLI components for improved organization
This commit renames `prompts.py` to `commit_prompts.py` and `cli.py` to `commit.py` in the `git_acp/cli` package. The changes aim to improve the separation of concerns by making the CLI subpackage more focused on its primary function, which is generating commit messages with AI. This refactoring enhances code readability and maintainability, making it easier to understand and modify individual components separately.
