+++ git_acp/cli/__init__.py
@@ -1,6 +1,6 @@
 """CLI package for git-acp."""
 
-from .cli import main
+from .commit import main
 from .pr import pr
 
 __all__ = ['main', 'pr']
